refactor cctp v1 and v2
